feat: migrate to Nuxt 3 - fixing plugins
