+++ package.json
@@ -9,20 +9,22 @@
     "postinstall": "nuxt prepare"
   },
   "devDependencies": {
+    "@types/js-cookie": "^3.0.3",
     "@types/lodash.debounce": "^4.0.7",
     "@types/mapbox-gl": "^2.7.10",
     "@types/node": "^18.14.1",
+    "ajv": "^6.10.1",
     "csv": "^6.2.7",
     "got": "^9.6.0",
-    "slugify": "^1.3.4",
-    "ramda": "^0.26.1",
-    "ajv": "^6.10.1",
     "nuxt": "^3.2.2",
+    "ramda": "^0.26.1",
+    "slugify": "^1.3.4",
     "vue-i18n": "^9.2.2"
   },
   "dependencies": {
     "@pinia/nuxt": "^0.4.7",
     "floating-vue": "^2.0.0-beta.20",
+    "js-cookie": "^2.2.1",
     "lodash.debounce": "^4.0.8",
     "mapbox-gl": "^2.13.0",
     "normalize.css": "^8.0.1",
