+++ src/plugins/persist-state.client.js
@@ -2,9 +2,10 @@
  * Persist and rehydrate app state between page loads
  * @see https://github.com/robinvdvleuten/vuex-persistedstate#nuxtjs
  */
-import createPersistedState from 'vuex-persistedstate'
+//import createPersistedState from 'vuex-persistedstate'
 
-export default ({ store }) => {
+export default defineNuxtPlugin((app) => {
+/*export default ({ store }) => {
   createPersistedState({
     key: 'app-state',
     paths: [
@@ -13,4 +14,5 @@ export default ({ store }) => {
       'hasSeenOnboarding',
     ],
   })(store)
-}
+}*/
+});
